BUG: Introduce salted passwords.
The passwords was not properly protected as they did not use a salt
mechanism and only used weak hashing algorithms. Switching to the
password library makes attacks harde, not impossible to do but harder
to succeed.

The example config uses the same salt as the default admin users is
installed with. Please note that changing the salt in the config file
will render the default admin user password useless.

closes: #19