BUG: Corrected syntax for getMembers().
The syntax for getMembers() was to send true to deliver the address for
the members. The report creation page was set to send the current date
and that would make the address part true because PHP read it as being
not zero.

Also added triple equal to make sure it was a boolean sent.