BUG: Removed exact match and leading string match for find functions.
The find functions did first a search for an exact match of the phrase,
in cases of names this was insane because there can certainly be more
than one "Martin" in the database.
We also did search for all records starting with the string and for
records containing the string if no exact match was found.

By removing exact match and leading match search and just keeping the
contains string match the result will be more natural and retains most
functionality.

closes #1018