BUG: Get current members in report is wrong
This is due to multiple periods overlapping oneanother.
What I've done is that I've made sure that there will be only info regarding
the person and not the period type in the report (as stated in the name).
Also, there will be an group by on the SSN just to make sure that
we avoid duplicate entrys for members.