FEATURE: Add database internal versioning and upgrade path.
Adding a settings table and setting initial database version to 0.

Future database upgrade will be placed in db/upgrade/number.db. The
application will find the upgrade files and execute them in order and
make the database upgraded.

A point to remeber is to check if the upgrade is needed, someone can
have applied the change on before hand for whatever reason and they
should not have duplicate entries and what not.