BUG: Set initial database version to 0.
When importing the initial database from db.sql as described in
README the settings table is installed but no version for DB
inserted. Makes the application halt (by design).

Adding the initial version 0 string to the base
definition. This "must" be bumped when doing database updates in
the future, to avoid executing the upgrade scripts on new setups
- they can be correct from the beginning.

Refs: #1246