FEATURE: Add DB helper 'tableexists(name)'.
Added a DB helper library file db/functions.php with a single function
named 'tableexitsts' that takes exactly one option and that's the name
of the table to look for. If it exists in the current database true is
returned otherwise false is returned.

Extended the PDO Interface class with a method to retrieve the
database name in use as that was needed to complete the table lookup
query.