FEATURE: Add payment information to report.
During some refactoring the payment information was lost in the report
part.
Extended the getMembers() function to honor the payment setting that
was sent to it and actually retrieve the active payment for the
members.

Took the opportunity to add " as devider between strings for easier
parsing of names, addresses and so on.