Created the ability to log events, Refs #7
These changes enables logging together with the creation of and
id column for the api table.

Logs are saved as an JSON-string in the database, the JSON contains the userid of the
local- or apiuser that used the function, the usertype (currently local or api) as well
as the timestamp of the event together with the custom message.