BUG: Default value for wrong address must be 0 (or 1).
When creating a new user and not selecting the checkbox for "wrong
address" the value was set to NULL. Only 0 or 1 are acceptable values
and thus this produced a failure. As adding a new member by hand must
always be considered having the correct address it's more convinient
to just assume that the address stated is correct and not make a fuzz
about it. If they do click the checkbox it will have 1 as value and
all is still well.