REFACTOR: Use CSS to select background color and hover color for table rows.
Instead of using a hardcoded value inside the view to paint the
background color of a table layout with alternating background color
per odd and even lines the color setting was moved to CSS.

At the same time removed the javascript lines for showing different
background color of the hovered lines using javascript. CSS have
support for that as well.