FEATURE: Add columnexists() check for DB upgrade script.
When trying to add a extra column that is already present a fatal error
will be presented instead of a silent pass.
Add a new funciton to db/functions.php that take a table name and a
column name and returns true if the queried combination exists. Useful
in upgrading when you do not want to add a already present column name.