BUG: Use ID as base for membershiptype control in addPayment (via API).
This is a strange construct, you check that the supplied membership type
is valid by selecting it using the ID and such things.

On the other hand you can not fetc the ID from the API, you need to know
it and harcode it into the upplying application. See #27 for more on
that note.